Make mock-service-endpoint a workspace library
