+++ tools/mock-service-endpoint/Cargo.toml
@@ -22,8 +22,10 @@ prost = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
-tokio-stream = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 thiserror = { workspace = true }
 workspace-hack = { version = "0.1", path = "../../workspace-hack" }
+
+[lib]
+path = "src/lib.rs"
